Moving 'Error' Handlers To Separate File
