+++ src/api/handler/handlers.go
@@ -5,29 +5,6 @@ import (
 	"net/http"
 )
 
-// Error represents a handler error. It provides methods for a HTTP status
-// code and embeds the built-in error interface.
-type Error interface {
-	error
-	Status() int
-}
-
-// StatusError represents an error with an associated HTTP status code.
-type StatusError struct {
-	Code int
-	Err  error
-}
-
-// Error gets the error string of the error
-func (se StatusError) Error() string {
-	return se.Err.Error()
-}
-
-// Status gets the status code of the error
-func (se StatusError) Status() int {
-	return se.Code
-}
-
 // Signature for a function that includes http.Handler + env parameters & returns an error-type
 type mp_envhandler_fn func(rw http.ResponseWriter, req *http.Request, env *Env) error
